change word type declaration to ensure 4-byte type (allows working on 64-bit systems)

+++ firmimgtool.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <time.h>
 
-typedef unsigned long  word;
+typedef int32_t  word;
 typedef unsigned short half;
 typedef unsigned char  byte;
